Correctly set initial camera angle.
According to Joey de Vries on LeanOpenGL, the GLFW mouse callback function is
called on application startup when the window first gets focus. This behavior
is inconsistent for me, happening on sometimes. When it does not happen, the
camera jerks the first time the mouse is moved. This can be solved by setting
the initial value of camera_front to the value it will become once set in the
mouse callback function. The mouse callback function sets it according to the
initial values of yaw and pitch.